+++ CHANGELOG.md
@@ -16,7 +16,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- The filters pane in course search is now a sliding drawer on mobile.
+- When displaying related objects on a category, respect page tree ordering,
+- The filters pane in course search is now a sliding drawer on mobile,
 - Use SVG icons instead of an icon font.
 
 ### Fixed