+++ CHANGELOG.md
@@ -29,6 +29,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Fix add plugin to team and organization on fragment_course_content template
   when they are empty.
 
+### Removed
+
+- Anti-pattern in factories to create related objects from the targeted page.
+
 ## [1.9.2] - 2019-09-20
 
 ### Fixed